id changes

+++ src/main/resources/db/migration/V1__create_tables.sql
@@ -8,8 +8,8 @@ CREATE TABLE IF NOT EXISTS player(
 CREATE TABLE IF NOT EXISTS game(
     id bigint GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     uuid VARCHAR(36) NOT NULL,
-    white_player_id integer,
-    black_player_id integer,
+    white_player_id bigint,
+    black_player_id bigint,
     status VARCHAR(48),
     FOREIGN KEY (white_player_id) REFERENCES player(id),
     FOREIGN KEY (black_player_id) REFERENCES player(id)
@@ -22,7 +22,7 @@ CREATE TABLE IF NOT EXISTS piece(
     status VARCHAR(8) NOT NULL,
     position_row integer NOT NULL,
     position_col integer NOT NULL,
-    game_id integer NOT NULL,
+    game_id bigint NOT NULL,
     FOREIGN KEY (game_id) REFERENCES game(id)
 );
 
@@ -32,9 +32,9 @@ CREATE TABLE IF NOT EXISTS move(
     src_row integer NOT NULL,
     dst_col integer NOT NULL,
     dst_row integer NOT NULL,
-    moving_piece_id integer NOT NULL,
+    moving_piece_id bigint NOT NULL,
     move_type VARCHAR(32) NOT NULL,
-    taken_piece_id integer,
+    taken_piece_id bigint,
     FOREIGN KEY (moving_piece_id) REFERENCES piece(id),
     FOREIGN KEY (taken_piece_id) REFERENCES piece(id)
 );
